requirements.txt
dj-database-url==0.4.1
django-sortedm2m==1.3.2
Django==1.9.7
flake8==2.6.2
gunicorn==19.6.0
psycopg2==2.6.1
pytest-django==2.9.1
pytest-factoryboy==1.2.1
sendgrid==2.2.1
whitenoise==3.2
